complete anonymous page

+++ include/vm/vm.h
@@ -125,7 +125,7 @@ void vm_dealloc_page (struct page *page);
 bool vm_claim_page (void *va);
 enum vm_type page_get_type (struct page *page);
 
-uint64_t page_hash (const struct hash_elem *p_, void *aux UNUSED);
+unsigned page_hash (const struct hash_elem *p_, void *aux UNUSED);
 bool page_less (const struct hash_elem *a_, const struct hash_elem *b_, void *aux UNUSED);
 void hash_action (struct hash_elem *e, void *aux);
 #endif  /* VM_VM_H */
